added cpu folder, fixed absolute_write_x
