+++ emulator/src/cpu/cpu.rs
@@ -1,7 +1,7 @@
-use crate::{
-    bus::{Bus, BusLike},
-    enums::Mirroring,
-};
+use crate::bus::BusLike;
+use crate::cpu::micro_instructions::{MicroInstruction, MicroInstructionSequence};
+use crate::cpu::operations::Operation;
+use crate::cpu::registers::Registers;
 
 pub struct CPU<T: BusLike> {
     bus: T,
@@ -11,31 +11,8 @@ pub struct CPU<T: BusLike> {
     current_micro_instruction: Option<MicroInstruction>,
 }
 
-pub struct Registers {
-    x: u8,
-    y: u8,
-    a: u8,
-    program_counter: u16,
-    stack_ptr: u8,
-    status: u8,
-    operation: u8,
-    adl: u8,
-    adh: u8,
-    bal: u8,
-    bah: u8,
-    ial: u8,
-    decoded_addressing_mode: Option<MicroInstructionSequence>,
-    decoded_operation: Option<MicroInstructionSequence>,
-    memory_buffer: u8,
-}
-
-pub struct MicroInstructionSequence {
-    sequence: Vec<MicroInstruction>,
-    idx: usize,
-}
-
 #[derive(Clone, PartialEq, Debug)]
-enum CPUFlag {
+pub enum CPUFlag {
     CarryBit,
     Zero,
     InterruptDisable,
@@ -46,791 +23,10 @@ enum CPUFlag {
     Negative,
 }
 
-#[derive(Clone, PartialEq, Debug)]
-enum MicroInstruction {
-    Empty,
-    ReadOperationCode,
-    DecodeOperation,
-
-    ImmediateRead,
-    ReadAdl,
-    ReadAdh,
-    ReadZeroPage,
-    ReadAbsolute,
-    ReadBal,
-    ReadBah,
-    ReadAdlIndirectBal,
-    ReadAdhIndirectBal,
-    ReadZeroPageBalX,
-    ReadZeroPageBalY,
-    ReadAdlAdhAbsoluteX,
-    ReadAdlAdhAbsoluteY,
-    ReadIal,
-    ReadBalIndirectIal,
-    ReadBahIndirectIal,
-
-    WriteZeroPage,
-    WriteAbsolute,
-    WriteZeroPageBalX,
-    WriteAbsoluteX,
-
-    ShiftLeftAccumulator,
-    ShiftLeftMemoryBuffer,
-
-    IncrementMemoryBuffer,
-    IncrementX,
-    IncrementY,
-    DecrementMemoryBuffer,
-    DecrementX,
-    DecrementY,
-
-    LoadAccumulator,
-    LoadX,
-    LoadY,
-
-    And,
-}
-
 #[derive(PartialEq, Debug)]
-enum CPUState {
+pub enum CPUState {
     Fetching,
     Execution,
-}
-
-#[derive(PartialEq, Debug)]
-enum Operation {
-    AslA,
-    AslZeroPage,
-    AslZeroPageX,
-    AslAbsolute,
-    IncMemZeroPage,
-    IncMemZeroPageX,
-    IncMemAbsolute,
-    IncMemAbsoluteX,
-    IncX,
-    IncY,
-    DecMemZeroPage,
-    DecMemZeroPageX,
-    DecMemAbsolute,
-    DecMemAbsoluteX,
-    DecX,
-    DecY,
-    LoadAccImm,
-    LoadAccZeroPage,
-    LoadAccZeroPageX,
-    LoadAccAbsolute,
-    LoadAccAbsoluteX,
-    LoadAccAbsoluteY,
-    LoadAccIndirectX,
-    LoadAccIndirectY,
-    LoadXImm,
-    LoadXZeroPage,
-    LoadXZeroPageY,
-    LoadXAbsolute,
-    LoadXAbsoluteY,
-    LoadYImm,
-    LoadYZeroPage,
-    LoadYZeroPageX,
-    LoadYAbsolute,
-    LoadYAbsoluteX,
-    AndImm,
-    AndZeroPage,
-    AndZeroPageX,
-    AndAbsolute,
-    AndAbsoluteX,
-    AndAbsoluteY,
-    AndIndirectX,
-    AndIndirectY,
-}
-
-struct OperationMicroInstructions {
-    pub addressing_sequence: Option<MicroInstructionSequence>,
-    pub operation_sequence: MicroInstructionSequence,
-}
-
-impl Operation {
-    fn get_micro_instructions(&self) -> OperationMicroInstructions {
-        let zero_page_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadAdl,
-            MicroInstruction::ReadZeroPage,
-        ]);
-        let zero_page_x_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadBal,
-            MicroInstruction::Empty, // Because we can add it in the next step easily
-            MicroInstruction::ReadZeroPageBalX,
-        ]);
-        let zero_page_y_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadBal,
-            MicroInstruction::Empty,
-            MicroInstruction::ReadZeroPageBalY,
-        ]);
-        let absolute_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadAdl,
-            MicroInstruction::ReadAdh,
-            MicroInstruction::ReadAbsolute,
-        ]);
-        let indirect_x_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadBal,
-            MicroInstruction::Empty, // Because we can add it in the next step easily
-            MicroInstruction::ReadAdlIndirectBal,
-            MicroInstruction::ReadAdhIndirectBal,
-            MicroInstruction::ReadAbsolute,
-        ]);
-        let absolute_x_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadBal,
-            MicroInstruction::ReadBah,
-            MicroInstruction::ReadAdlAdhAbsoluteX,
-            // TODO: Check if this is correct (T4 is optional if page boundary is not crossed)
-        ]);
-        let absolute_y_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadBal,
-            MicroInstruction::ReadBah,
-            MicroInstruction::ReadAdlAdhAbsoluteY,
-        ]);
-        let indirect_y_addressing = MicroInstructionSequence::new(vec![
-            MicroInstruction::ReadIal,
-            MicroInstruction::ReadBalIndirectIal,
-            MicroInstruction::ReadBahIndirectIal,
-            MicroInstruction::ReadAdlAdhAbsoluteY,
-            // TODO: Same as absolute_x_addressing
-        ]);
-        let immediate_addressing =
-            MicroInstructionSequence::new(vec![MicroInstruction::ImmediateRead]);
-
-        match self {
-            Self::AslA => OperationMicroInstructions {
-                addressing_sequence: None,
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::ShiftLeftAccumulator,
-                ]),
-            },
-            Self::AslZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::ShiftLeftMemoryBuffer,
-                    MicroInstruction::WriteZeroPage,
-                ]),
-            },
-            Self::AslZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::ShiftLeftMemoryBuffer,
-                    MicroInstruction::WriteZeroPageBalX,
-                ]),
-            },
-            Self::AslAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::ShiftLeftMemoryBuffer,
-                    MicroInstruction::WriteAbsolute,
-                ]),
-            },
-            Self::IncMemZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementMemoryBuffer,
-                    MicroInstruction::WriteZeroPage,
-                ]),
-            },
-            Self::IncMemZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementMemoryBuffer,
-                    MicroInstruction::WriteZeroPageBalX,
-                ]),
-            },
-            Self::IncMemAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementMemoryBuffer,
-                    MicroInstruction::WriteAbsolute,
-                ]),
-            },
-            Self::IncMemAbsoluteX => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementMemoryBuffer,
-                    MicroInstruction::WriteAbsoluteX,
-                ]),
-            },
-            Self::IncX => OperationMicroInstructions {
-                addressing_sequence: None,
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementX,
-                ]),
-            },
-            Self::IncY => OperationMicroInstructions {
-                addressing_sequence: None,
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::IncrementY,
-                ]),
-            },
-            Self::DecMemZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementMemoryBuffer,
-                    MicroInstruction::WriteZeroPage,
-                ]),
-            },
-            Self::DecMemZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementMemoryBuffer,
-                    MicroInstruction::WriteZeroPageBalX,
-                ]),
-            },
-            Self::DecMemAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementMemoryBuffer,
-                    MicroInstruction::WriteAbsolute,
-                ]),
-            },
-            Self::DecMemAbsoluteX => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementMemoryBuffer,
-                    MicroInstruction::WriteAbsoluteX,
-                ]),
-            },
-            Self::DecX => OperationMicroInstructions {
-                addressing_sequence: None,
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementX,
-                ]),
-            },
-            Self::DecY => OperationMicroInstructions {
-                addressing_sequence: None,
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::DecrementY,
-                ]),
-            },
-            Self::LoadAccImm => OperationMicroInstructions {
-                addressing_sequence: Some(immediate_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccAbsoluteX => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccAbsoluteY => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccIndirectX => OperationMicroInstructions {
-                addressing_sequence: Some(indirect_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadAccIndirectY => OperationMicroInstructions {
-                addressing_sequence: Some(indirect_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![
-                    MicroInstruction::LoadAccumulator,
-                ]),
-            },
-            Self::LoadXImm => OperationMicroInstructions {
-                addressing_sequence: Some(immediate_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadX]),
-            },
-            Self::LoadXZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadX]),
-            },
-            Self::LoadXZeroPageY => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadX]),
-            },
-            Self::LoadXAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadX]),
-            },
-            Self::LoadXAbsoluteY => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadX]),
-            },
-            Self::LoadYImm => OperationMicroInstructions {
-                addressing_sequence: Some(immediate_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadY]),
-            },
-            Self::LoadYZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadY]),
-            },
-            Self::LoadYZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadY]),
-            },
-            Self::LoadYAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadY]),
-            },
-            Self::LoadYAbsoluteX => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::LoadY]),
-            },
-            Self::AndImm => OperationMicroInstructions {
-                addressing_sequence: Some(immediate_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndZeroPage => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndZeroPageX => OperationMicroInstructions {
-                addressing_sequence: Some(zero_page_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndAbsolute => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndAbsoluteX => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndAbsoluteY => OperationMicroInstructions {
-                addressing_sequence: Some(absolute_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndIndirectX => OperationMicroInstructions {
-                addressing_sequence: Some(indirect_x_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-            Self::AndIndirectY => OperationMicroInstructions {
-                addressing_sequence: Some(indirect_y_addressing),
-                operation_sequence: MicroInstructionSequence::new(vec![MicroInstruction::And]),
-            },
-        }
-    }
-
-    fn get_opcode(&self) -> u8 {
-        match self {
-            Self::AslA => 0x0A,
-            Self::AslZeroPage => 0x06,
-            Self::AslZeroPageX => 0x16,
-            Self::AslAbsolute => 0x0E,
-            Self::IncMemZeroPage => 0xE6,
-            Self::IncMemZeroPageX => 0xF6,
-            Self::IncMemAbsolute => 0xEE,
-            Self::IncMemAbsoluteX => 0xFE,
-            Self::IncX => 0xE8,
-            Self::IncY => 0xC8,
-            Self::DecMemZeroPage => 0xC6,
-            Self::DecMemZeroPageX => 0xD6,
-            Self::DecMemAbsolute => 0xCE,
-            Self::DecMemAbsoluteX => 0xDE,
-            Self::DecX => 0xCA,
-            Self::DecY => 0x88,
-            Self::LoadAccImm => 0xA9,
-            Self::LoadAccZeroPage => 0xA5,
-            Self::LoadAccZeroPageX => 0xB5,
-            Self::LoadAccAbsolute => 0xAD,
-            Self::LoadAccAbsoluteX => 0xBD,
-            Self::LoadAccAbsoluteY => 0xB9,
-            Self::LoadAccIndirectX => 0xA1,
-            Self::LoadAccIndirectY => 0xB1,
-            Self::LoadXImm => 0xA2,
-            Self::LoadXZeroPage => 0xA6,
-            Self::LoadXZeroPageY => 0xB6,
-            Self::LoadXAbsolute => 0xAE,
-            Self::LoadXAbsoluteY => 0xBE,
-            Self::LoadYImm => 0xA0,
-            Self::LoadYZeroPage => 0xA4,
-            Self::LoadYZeroPageX => 0xB4,
-            Self::LoadYAbsolute => 0xAC,
-            Self::LoadYAbsoluteX => 0xBC,
-            Self::AndImm => 0x29,
-            Self::AndZeroPage => 0x25,
-            Self::AndZeroPageX => 0x35,
-            Self::AndAbsolute => 0x2D,
-            Self::AndAbsoluteX => 0x3D,
-            Self::AndAbsoluteY => 0x39,
-            Self::AndIndirectX => 0x21,
-            Self::AndIndirectY => 0x31,
-        }
-    }
-
-    fn get_operation(opcode: u8) -> Option<Self> {
-        match opcode {
-            0x0A => Some(Self::AslA),
-            0x06 => Some(Self::AslZeroPage),
-            0x16 => Some(Self::AslZeroPageX),
-            0x0E => Some(Self::AslAbsolute),
-            0xE6 => Some(Self::IncMemZeroPage),
-            0xF6 => Some(Self::IncMemZeroPageX),
-            0xEE => Some(Self::IncMemAbsolute),
-            0xFE => Some(Self::IncMemAbsoluteX),
-            0xE8 => Some(Self::IncX),
-            0xC8 => Some(Self::IncY),
-            0xC6 => Some(Self::DecMemZeroPage),
-            0xD6 => Some(Self::DecMemZeroPageX),
-            0xCE => Some(Self::DecMemAbsolute),
-            0xDE => Some(Self::DecMemAbsoluteX),
-            0xCA => Some(Self::DecX),
-            0x88 => Some(Self::DecY),
-            0xA9 => Some(Self::LoadAccImm),
-            0xA5 => Some(Self::LoadAccZeroPage),
-            0xB5 => Some(Self::LoadAccZeroPageX),
-            0xAD => Some(Self::LoadAccAbsolute),
-            0xBD => Some(Self::LoadAccAbsoluteX),
-            0xB9 => Some(Self::LoadAccAbsoluteY),
-            0xA1 => Some(Self::LoadAccIndirectX),
-            0xB1 => Some(Self::LoadAccIndirectY),
-            0xA2 => Some(Self::LoadXImm),
-            0xA6 => Some(Self::LoadXZeroPage),
-            0xB6 => Some(Self::LoadXZeroPageY),
-            0xAE => Some(Self::LoadXAbsolute),
-            0xBE => Some(Self::LoadXAbsoluteY),
-            0xA0 => Some(Self::LoadYImm),
-            0xA4 => Some(Self::LoadYZeroPage),
-            0xB4 => Some(Self::LoadYZeroPageX),
-            0xAC => Some(Self::LoadYAbsolute),
-            0xBC => Some(Self::LoadYAbsoluteX),
-            0x29 => Some(Self::AndImm),
-            0x25 => Some(Self::AndZeroPage),
-            0x35 => Some(Self::AndZeroPageX),
-            0x2D => Some(Self::AndAbsolute),
-            0x3D => Some(Self::AndAbsoluteX),
-            0x39 => Some(Self::AndAbsoluteY),
-            0x21 => Some(Self::AndIndirectX),
-            0x31 => Some(Self::AndIndirectY),
-            _ => None,
-        }
-    }
-}
-
-impl Registers {
-    fn new() -> Self {
-        Self {
-            x: 0x00,
-            y: 0x00,
-            a: 0x00,
-            program_counter: 0x0000,
-            stack_ptr: 0x00,
-            status: 0x00,
-            operation: 0x00,
-            adl: 0x00,
-            adh: 0x00,
-            bal: 0x00,
-            bah: 0x00,
-            ial: 0x00,
-            decoded_addressing_mode: None,
-            decoded_operation: None,
-            memory_buffer: 0x00,
-        }
-    }
-
-    fn get_operation(&mut self) -> &mut Option<MicroInstructionSequence> {
-        match self.decoded_addressing_mode {
-            Some(ref mut decoded_addressing_mode) => {
-                if decoded_addressing_mode.is_completed() {
-                    &mut self.decoded_operation
-                } else {
-                    &mut self.decoded_addressing_mode
-                }
-            }
-            None => &mut self.decoded_operation,
-        }
-    }
-
-    fn is_operation_completed(&self) -> bool {
-        match &self.decoded_operation {
-            Some(operation) => operation.is_completed(),
-            None => false,
-        }
-    }
-
-    fn set_flag(&mut self, flag: CPUFlag) {
-        self.status |= flag.value();
-    }
-
-    fn clear_flag(&mut self, flag: CPUFlag) {
-        self.status &= !flag.value();
-    }
-
-    fn set_flag_value(&mut self, flag: CPUFlag, value: bool) {
-        if value {
-            self.set_flag(flag);
-        } else {
-            self.clear_flag(flag);
-        }
-    }
-
-    fn is_flag_set(&self, flag: CPUFlag) -> bool {
-        self.status & flag.value() != 0
-    }
-
-    fn reset_flags(&mut self) {
-        self.status = 0x00;
-    }
-
-    fn step_program_counter(&mut self) {
-        self.program_counter += 1;
-    }
-
-    fn read_operation_code<T: BusLike>(&mut self, bus: &mut T) {
-        self.operation = bus.read(self.program_counter as u16);
-    }
-
-    fn decode_operation<T: BusLike>(&mut self, bus: &T) {
-        let operation_code = self.operation;
-        println!("Operation code: {:#X}", operation_code);
-
-        if let Some(operation) = Operation::get_operation(operation_code) {
-            let micro_instructions = operation.get_micro_instructions();
-            self.decoded_addressing_mode = micro_instructions.addressing_sequence;
-            self.decoded_operation = Some(micro_instructions.operation_sequence);
-        } else {
-            panic!("Operation not found for opcode: {:#X}", operation_code);
-        }
-
-        self.step_program_counter();
-    }
-
-    fn immediate_read<T: BusLike>(&mut self, bus: &mut T) {
-        self.memory_buffer = bus.read(self.program_counter);
-        self.step_program_counter();
-    }
-
-    fn read_adl<T: BusLike>(&mut self, bus: &mut T) {
-        self.adl = bus.read(self.program_counter);
-        self.step_program_counter();
-    }
-
-    fn read_adh<T: BusLike>(&mut self, bus: &mut T) {
-        self.adh = bus.read(self.program_counter);
-        self.step_program_counter();
-    }
-
-    fn read_zero_page<T: BusLike>(&mut self, bus: &mut T) {
-        println!("Reading zero page address: {:#X}", self.adl);
-        self.memory_buffer = bus.read(self.adl as u16);
-    }
-
-    fn read_absolute<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.adh as u16) << 8 | self.adl as u16;
-        self.memory_buffer = bus.read(address as u16);
-    }
-
-    fn read_bal<T: BusLike>(&mut self, bus: &mut T) {
-        self.bal = bus.read(self.program_counter as u16);
-        self.step_program_counter();
-    }
-
-    fn read_bah<T: BusLike>(&mut self, bus: &mut T) {
-        self.bah = bus.read(self.program_counter as u16);
-        self.step_program_counter();
-    }
-
-    fn read_adl_indirect_bal<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.bal + self.x) as usize;
-        self.adl = bus.read(address as u16);
-    }
-
-    fn read_adh_indirect_bal<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.bal + self.x + 1) as usize;
-        self.adh = bus.read(address as u16);
-    }
-
-    fn write_zero_page<T: BusLike>(&mut self, bus: &mut T) {
-        bus.write(self.adl as u16, self.memory_buffer);
-    }
-
-    fn write_absolute<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.adh as u16) << 8 | self.adl as u16;
-        bus.write(address as u16, self.memory_buffer);
-    }
-
-    fn write_absolute_x<T: BusLike>(&mut self, bus: &mut T) {
-        let address = ((self.adh as u16) << 8 | self.adl as u16) + self.x as u16;
-        bus.write(address, self.memory_buffer);
-    }
-
-    fn read_zero_page_bal_x<T: BusLike>(&mut self, bus: &mut T) {
-        // TODO: Be careful with overflow, check if it's correct
-
-        let address = (self.bal + self.x) as usize;
-        self.memory_buffer = bus.read(address as u16);
-    }
-
-    fn read_zero_page_bal_y<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.bal + self.y) as usize;
-        self.memory_buffer = bus.read(address as u16);
-    }
-
-    fn write_zero_page_bal_x<T: BusLike>(&mut self, bus: &mut T) {
-        let address = (self.bal + self.x) as usize;
-        bus.write(address as u16, self.memory_buffer);
-    }
-
-    fn read_adl_adh_absolute_index_register<T: BusLike>(
-        &mut self,
-        bus: &mut T,
-        index_register: u8,
-    ) {
-        let bal_address = self.bal as usize;
-        let bah_address = self.bah as usize;
-        let address = ((bah_address << 8) | bal_address) + (index_register as usize);
-        self.adh = self.bah;
-        self.adl = self.bal;
-        self.memory_buffer = bus.read(address as u16);
-    }
-
-    fn read_adl_adh_absolute_x<T: BusLike>(&mut self, bus: &mut T) {
-        self.read_adl_adh_absolute_index_register(bus, self.x);
-    }
-
-    fn read_adl_adh_absolute_y<T: BusLike>(&mut self, bus: &mut T) {
-        self.read_adl_adh_absolute_index_register(bus, self.y);
-    }
-
-    fn read_ial<T: BusLike>(&mut self, bus: &mut T) {
-        self.ial = bus.read(self.program_counter as u16);
-        self.step_program_counter();
-    }
-
-    fn read_bal_indirect_ial<T: BusLike>(&mut self, bus: &mut T) {
-        self.bal = bus.read(self.ial as u16);
-    }
-
-    fn read_bah_indirect_ial<T: BusLike>(&mut self, bus: &mut T) {
-        self.bah = bus.read(self.ial as u16 + 1);
-    }
-
-    fn shift_left_accumulator(&mut self) {
-        let is_carry = self.a & 0x80 != 0;
-        self.a <<= 1;
-        let is_negative = self.a & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::CarryBit, is_carry);
-        self.set_flag_value(CPUFlag::Zero, self.a == 0);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn shift_left_memory_buffer(&mut self) {
-        let is_carry = self.memory_buffer & 0x80 != 0;
-        self.memory_buffer <<= 1;
-        let is_negative = self.memory_buffer & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::CarryBit, is_carry);
-        self.set_flag_value(CPUFlag::Zero, self.memory_buffer == 0);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn increment_memory_buffer(&mut self) {
-        self.memory_buffer = self.memory_buffer.wrapping_add(1u8);
-        let is_zero = self.memory_buffer == 0;
-        let is_negative = self.memory_buffer & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn increment_x(&mut self) {
-        self.x = self.x.wrapping_add(1u8);
-        let is_zero = self.x == 0;
-        let is_negative = self.x & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn increment_y(&mut self) {
-        self.y = self.y.wrapping_add(1u8);
-        let is_zero = self.y == 0;
-        let is_negative = self.x & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn dec_memory_buffer(&mut self) {
-        self.memory_buffer = self.memory_buffer.wrapping_sub(1u8);
-        let is_zero = self.memory_buffer == 0;
-        let is_negative = self.memory_buffer & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn dec_x(&mut self) {
-        self.x = self.x.wrapping_sub(1u8);
-        let is_zero = self.x == 0;
-        let is_negative = self.x & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn dec_y(&mut self) {
-        self.y = self.y.wrapping_sub(1u8);
-        let is_zero = self.y == 0;
-        let is_negative = self.y & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn load_accumulator(&mut self) {
-        self.a = self.memory_buffer;
-        let is_zero = self.a == 0;
-        let is_negative = self.a & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn load_x(&mut self) {
-        self.x = self.memory_buffer;
-        let is_zero = self.x == 0;
-        let is_negative = self.x & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn load_y(&mut self) {
-        self.y = self.memory_buffer;
-        let is_zero = self.y == 0;
-        let is_negative = self.y & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
-
-    fn and(&mut self) {
-        self.a = self.a & self.memory_buffer;
-        let is_zero = self.a == 0;
-        let is_negative = self.a & 0x80 != 0;
-
-        self.set_flag_value(CPUFlag::Zero, is_zero);
-        self.set_flag_value(CPUFlag::Negative, is_negative);
-    }
 }
 
 impl<T: BusLike> CPU<T> {
@@ -939,9 +135,6 @@ impl<T: BusLike> CPU<T> {
             MicroInstruction::WriteZeroPageBalX => {
                 self.registers.write_zero_page_bal_x(&mut self.bus)
             }
-            MicroInstruction::WriteAbsoluteX => {
-                self.registers.write_absolute_x(&mut self.bus);
-            }
             MicroInstruction::ShiftLeftAccumulator => self.registers.shift_left_accumulator(),
             MicroInstruction::ShiftLeftMemoryBuffer => self.registers.shift_left_memory_buffer(),
             MicroInstruction::IncrementMemoryBuffer => self.registers.increment_memory_buffer(),
@@ -958,29 +151,8 @@ impl<T: BusLike> CPU<T> {
     }
 }
 
-impl MicroInstructionSequence {
-    fn new(sequence: Vec<MicroInstruction>) -> Self {
-        Self { sequence, idx: 0 }
-    }
-
-    fn get_micro_instruction(&self) -> &MicroInstruction {
-        &self.sequence[self.idx]
-    }
-
-    fn next(&mut self) {
-        self.idx += 1;
-    }
-
-    fn is_completed(&self) -> bool {
-        self.idx >= self.sequence.len()
-    }
-
-    fn reset(&mut self) {
-        self.idx = 0;
-    }
-}
 impl CPUFlag {
-    fn value(&self) -> u8 {
+    pub fn value(&self) -> u8 {
         match *self {
             Self::CarryBit => 1 << 0,
             Self::Zero => 1 << 1,
@@ -1524,7 +696,7 @@ mod tests {
         assert_eq!(cpu.state, CPUState::Fetching);
         assert_eq!(
             cpu.current_micro_instruction,
-            Some(MicroInstruction::WriteAbsoluteX)
+            Some(MicroInstruction::WriteAbsolute)
         );
 
         let read_value = cpu.bus.read(expected_address);
@@ -1736,7 +908,7 @@ mod tests {
         assert_eq!(cpu.state, CPUState::Fetching);
         assert_eq!(
             cpu.current_micro_instruction,
-            Some(MicroInstruction::WriteAbsoluteX)
+            Some(MicroInstruction::WriteAbsolute)
         );
 
         let read_value = cpu.bus.read(expected_address);
