+++ emulator/src/cpu/mod.rs
@@ -1,0 +1,4 @@
+pub mod cpu;
+pub mod micro_instructions;
+pub mod operations;
+pub mod registers;
